+++ Editor/Loading/SceneLoadingSettingsProvider.cs
@@ -1,0 +1,20 @@
+using UnityEditor;
+using ActionCode.ScriptableSettingsProvider.Editor;
+
+namespace ActionCode.SceneManagement.Editor
+{
+    public sealed class SceneLoadingSettingsProvider :
+        AbstractScriptableSettingsProvider<SceneLoadingSettings>
+    {
+        public SceneLoadingSettingsProvider() :
+            base("ActionCode/Scene Loading")
+        { }
+
+        protected override string GetConfigName() =>
+            "com.actioncode.scene-management";
+
+        [SettingsProvider]
+        private static SettingsProvider CreateProjectSettingsMenu() =>
+            new SceneLoadingSettingsProvider();
+    }
+}